Adding the Playbook and the events (I'm calling them plays. Sue me) that I can easily parse. Including:  Drawing a Walk, Fielder's Choice (buggy?), Home Runs (solo and loaded), Runs (and scores), Sacrifice (not getting flys?), Stealing Bases  Batter on Plate, Bottom of Inning Changeover, Top of Inning Changeover  Game Start, Game End  Caught Stealing, Flyouts, Groundouts (TDHAGOTET NEVER FORGET), Strikeout  Ball, Foul Ball, Strike
All of these plays are subclasses of an abstract class which expects a regex, and matches it against the lastUpdate parameter, and parses the information it can from that string and from the state of the game passed to it.
These will form the core of the watching experience. SampleScene has an illustration of the game being played with last season's Firefighters/Mills Playoff match